requirements.txt
pydantic
typing
typing-extensions
kubernetes
requests
dateparser
fastapi
uvicorn
aiofiles
starlette_exporter
